requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
motor==3.3.2
pymongo==4.6.0
alembic==1.13.1

# Data Processing and Analysis
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.4.0

# Visualization
plotly==5.17.0
kaleido==0.2.1
tenacity

# Machine Learning
xgboost==2.0.3
lightgbm==4.3.0
catboost==1.2.2
joblib==1.3.2
optuna==3.5.0

# MLOps and Model Versioning
mlflow==2.9.2
dvc[azure]==3.48.4

# Configuration and Utilities
pydantic-settings==2.1.0
python-dotenv==1.0.0
dvc[s3]==3.48.4

# Database - PostgreSQL
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Database - MongoDB
pymongo==4.6.0
motor==3.3.2

# Configuration and Environment
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# File Processing
openpyxl==3.1.2
xlrd==2.0.1
pyarrow==14.0.1

# Model Deployment and Containerization
docker<7.0.0
kubernetes==28.1.0
pyyaml==6.0.1
jinja2==3.1.2

# API and Monitoring
prometheus-client==0.19.0
gunicorn==21.2.0

# Development and Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
